fix: remove redundant accesible element for better usability
