+++ src/screens/Departures/components/QuaySection.tsx
@@ -5,12 +5,21 @@ import {ActivityIndicator, View} from 'react-native';
 import {FlatList} from 'react-native-gesture-handler';
 import ThemeText from '@atb/components/text';
 import ThemeIcon from '@atb/components/theme-icon/theme-icon';
-import {useTranslation} from '@atb/translations';
+import {
+  CancelledDepartureTexts,
+  dictionary,
+  useTranslation,
+} from '@atb/translations';
 import {ExpandMore, ExpandLess} from '@atb/assets/svg/mono-icons/navigation';
 import {EstimatedCall, Quay} from '@atb/api/types/departures';
 import DeparturesTexts from '@atb/translations/screens/Departures';
 import EstimatedCallItem from './EstimatedCallItem';
 import SectionSeparator from '@atb/components/sections/section-separator';
+import {
+  formatToClockOrLongRelativeMinutes,
+  formatToSimpleDate,
+} from '@atb/utils/date';
+import {isToday, parseISO} from 'date-fns';
 
 type QuaySectionProps = {
   quay: Quay;
@@ -41,7 +50,7 @@ export default function QuaySection({
   const [isHidden, setIsHidden] = useState(false);
   const styles = useStyles();
   const departures = getDeparturesForQuay(data, quay);
-  const {t} = useTranslation();
+  const {t, language} = useTranslation();
 
   return (
     <View testID={testID}>
@@ -105,7 +114,7 @@ export default function QuaySection({
                       departure.cancellation,
                     );
                 }}
-                accessibilityHint={t(DeparturesTexts.a11yEstimatedCallItemHint)}
+                accessibilityLabel={getA11yLineLabel(departure, t, language)}
               >
                 <EstimatedCallItem
                   departure={departure}
@@ -158,6 +167,38 @@ export default function QuaySection({
       </Sections.Section>
     </View>
   );
+}
+
+function getA11yLineLabel(departure: any, t: any, language: any) {
+  const line = departure.serviceJourney?.line;
+  const a11yLine = line?.publicCode
+    ? `${t(DeparturesTexts.line)} ${line?.publicCode},`
+    : '';
+  const a11yFrontText = departure.destinationDisplay?.frontText
+    ? `${departure.destinationDisplay?.frontText}.`
+    : '';
+
+  let a11yDateInfo = '';
+  if (departure.expectedDepartureTime) {
+    const a11yClock = formatToClockOrLongRelativeMinutes(
+      departure.expectedDepartureTime,
+      language,
+      t(dictionary.date.units.now),
+      9,
+    );
+    const a11yTimeWithRealtimePrefix = departure.realtime
+      ? a11yClock
+      : t(dictionary.a11yMissingRealTimePrefix) + a11yClock;
+    const parsedDepartureTime = parseISO(departure.expectedDepartureTime);
+    const a11yDate = !isToday(parsedDepartureTime)
+      ? formatToSimpleDate(parsedDepartureTime, language) + ','
+      : '';
+    a11yDateInfo = `${a11yDate} ${a11yTimeWithRealtimePrefix}`;
+  }
+
+  return `${
+    departure.cancellation && t(CancelledDepartureTexts.message)
+  } ${a11yLine} ${a11yFrontText} ${a11yDateInfo}`;
 }
 
 function getDeparturesForQuay(
